backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
pillow
pytesseract
PyPDF2
python-docx 
sqlalchemy
passlib[bcrypt]
python-jose
httpx